Move keepalive_spec.lua to test directory
